small splashes and bubbles
Fixed up the small splashes from water droplets, and added the plunge bubbles. Shallow water running around and deep water jumping looks so good.

Also reordered some methods in Mario.js according to how they’re ordered in sm64 and added set_mario_action_submerged.
