+++ src/game/behaviors/water_objs.inc.js
@@ -7,6 +7,7 @@ import { oPosX, oPosY, oPosZ,
          oWaterObjUnkFC,
          oWaterObjUnk100,
          oInteractStatus,
+         oTimer,
          ACTIVE_FLAG_DEACTIVATED                } from "../../include/object_constants"
 import { INT_STATUS_INTERACTED                  } from "../Interaction"
 import { obj_mark_for_deletion,
@@ -150,12 +151,14 @@ export const bhv_small_bubbles_loop = () => {
 }
 
 export const bhv_fish_group_loop = () => {
+    const o = ObjectListProc.gCurrentObject
     if (ObjectListProc.gMarioCurrentRoom == 15 || ObjectListProc.gMarioCurrentRoom == 7)
         if (window.gGlobalTimer & 1)
             spawn_object(o, MODEL_WHITE_PARTICLE_SMALL, bhvSmallParticleBubbles);
 }
 
 export const bhv_water_waves_init= () => {
+    const o = ObjectListProc.gCurrentObject
     for (let i = 0; i < 3; i++) {
         spawn_object(o, MODEL_WHITE_PARTICLE_SMALL, bhvSmallParticle);
     }
